+++ src/main/java/io/supertokens/session/Session.java
@@ -49,6 +49,8 @@ import io.supertokens.session.info.TokenInfo;
 import io.supertokens.session.jwt.JWT;
 import io.supertokens.session.refreshToken.RefreshToken;
 import io.supertokens.storageLayer.StorageLayer;
+import io.supertokens.useridmapping.UserIdMapping;
+import io.supertokens.useridmapping.UserIdType;
 import io.supertokens.utils.Utils;
 import org.jetbrains.annotations.TestOnly;
 
@@ -139,11 +141,28 @@ public class Session {
         }
 
         String primaryUserId = recipeUserId;
-        if (storage.getType().equals(STORAGE_TYPE.SQL)) {
+
+        if (storage.getType() == STORAGE_TYPE.SQL) {
+            io.supertokens.pluginInterface.useridmapping.UserIdMapping userIdMapping = UserIdMapping.getUserIdMapping(
+                    tenantIdentifier.toAppIdentifier(), storage, recipeUserId, UserIdType.EXTERNAL);
+            if (userIdMapping != null) {
+                recipeUserId = userIdMapping.superTokensUserId;
+            }
+
             primaryUserId = StorageUtils.getAuthRecipeStorage(storage)
                     .getPrimaryUserIdStrForUserId(tenantIdentifier.toAppIdentifier(), recipeUserId);
             if (primaryUserId == null) {
                 primaryUserId = recipeUserId;
+            }
+
+            HashMap<String, String> userIdMappings = UserIdMapping.getUserIdMappingForSuperTokensUserIds(
+                    tenantIdentifier.toAppIdentifier(), storage,
+                    new ArrayList<>(Arrays.asList(primaryUserId, recipeUserId)));
+            if (userIdMappings.containsKey(primaryUserId)) {
+                primaryUserId = userIdMappings.get(primaryUserId);
+            }
+            if (userIdMappings.containsKey(recipeUserId)) {
+                recipeUserId = userIdMappings.get(recipeUserId);
             }
         }